+++ src/texture.cc
@@ -23,7 +23,7 @@ vec4 texture::sample(vec2 uv) {
     return getColor(x, y);
 }
 
-vec4 texture::getColor(float x, float y) {
+vec4 texture::getColor(int x, int y) {
     vec4 ret = {0.f,0.f,0.f,1.f};
     if(!data || x < 0 || y < 0 || x >= width || y >= height)
         return ret;
